+++ src/test/java/org/jabref/model/groups/AutomaticKeywordGroupTest.java
@@ -10,12 +10,23 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class AutomaticKeywordGroupTest {
+class AutomaticKeywordGroupTest {
 
     @Test
-    public void createSubgroupsForTwoKeywords() throws Exception {
+    void createSubgroupsForTwoKeywords() throws Exception {
         AutomaticKeywordGroup keywordsGroup = new AutomaticKeywordGroup("Keywords", GroupHierarchyType.INDEPENDENT, StandardField.KEYWORDS, ',', '>');
         BibEntry entry = new BibEntry().withField(StandardField.KEYWORDS, "A, B");
+
+        Set<GroupTreeNode> expected = new HashSet<>();
+        expected.add(GroupTreeNode.fromGroup(new WordKeywordGroup("A", GroupHierarchyType.INCLUDING, StandardField.KEYWORDS, "A", true, ',', true)));
+        expected.add(GroupTreeNode.fromGroup(new WordKeywordGroup("B", GroupHierarchyType.INCLUDING, StandardField.KEYWORDS, "B", true, ',', true)));
+        assertEquals(expected, keywordsGroup.createSubgroups(entry));
+    }
+
+    @Test
+    void createSubgroupsIgnoresEmptyKeyword() throws Exception {
+        AutomaticKeywordGroup keywordsGroup = new AutomaticKeywordGroup("Keywords", GroupHierarchyType.INDEPENDENT, StandardField.KEYWORDS, ',', '>');
+        BibEntry entry = new BibEntry().withField(StandardField.KEYWORDS, "A, ,B");
 
         Set<GroupTreeNode> expected = new HashSet<>();
         expected.add(GroupTreeNode.fromGroup(new WordKeywordGroup("A", GroupHierarchyType.INCLUDING, StandardField.KEYWORDS, "A", true, ',', true)));